+++ packages/ses/src/assert.js
@@ -1,5 +1,0 @@
-export function assert(condition, errorMessage) {
-  if (!condition) {
-    throw new TypeError(errorMessage);
-  }
-}