+++ packages/ses/test262/tame-global-error-object.js
@@ -1,5 +1,5 @@
 import test262Runner from '@agoric/test262-runner';
-import tameErrorConstructor from '../src/tame-error-constructor.js';
+import tameErrorConstructor from '../src/error/tame-error-constructor.js';
 
 test262Runner({
   testDirs: ['/test/built-ins/Error'],